# mcp_server/openapi.yaml
openapi: 3.1.0
info:
  title: Aurora MCP Server
  description: Model Context Protocol Server for AI Agents - Provides filesystem and process tools for the Aurora-X Ultra codebase
  version: 1.0.0
servers:
  - url: https://993461dc-3757-4f77-8c01-8bec9bddf5ee-00-3su7j9cyl1vd5.picard.replit.dev
    description: Production server
paths:
  /fs/read:
    post:
      operationId: readFile
      summary: Read file contents
      description: Read the contents of a file from the Aurora codebase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - path
              properties:
                path:
                  type: string
                  description: Path to the file to read (relative to workspace root)
      responses:
        '200':
          description: File contents
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: string
                    description: The file contents
                  path:
                    type: string
                    description: The resolved file path
        '400':
          description: Error reading file
  /fs/write:
    post:
      operationId: writeFile
      summary: Write to a file
      description: Write content to a file in the Aurora codebase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - path
                - content
              properties:
                path:
                  type: string
                  description: Path to the file to write
                content:
                  type: string
                  description: Content to write to the file
      responses:
        '200':
          description: Write successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  path:
                    type: string
                  bytes_written:
                    type: integer
        '400':
          description: Error writing file
  /fs/list:
    post:
      operationId: listDirectory
      summary: List directory contents
      description: List files and directories in a given path
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                path:
                  type: string
                  description: Directory path to list
                  default: "."
                recursive:
                  type: boolean
                  description: Whether to list recursively
                  default: false
      responses:
        '200':
          description: Directory listing
          content:
            application/json:
              schema:
                type: object
                properties:
                  entries:
                    type: array
                    items:
                      type: object
                      properties:
                        path:
                          type: string
                        type:
                          type: string
                          enum: [file, directory]
                        size:
                          type: integer
                  count:
                    type: integer
        '400':
          description: Error listing directory
  /process/run:
    post:
      operationId: runCommand
      summary: Run a shell command
      description: Execute a shell command and return the output
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - command
              properties:
                command:
                  type: string
                  description: Shell command to execute
                timeout:
                  type: integer
                  description: Timeout in seconds
                  default: 30
      responses:
        '200':
          description: Command output
          content:
            application/json:
              schema:
                type: object
                properties:
                  stdout:
                    type: string
                    description: Standard output
                  stderr:
                    type: string
                    description: Standard error
                  exit_code:
                    type: integer
                    description: Exit code
        '400':
          description: Command error
  /process/exec:
    post:
      operationId: execProgram
      summary: Execute a program
      description: Execute a program with arguments
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - program
              properties:
                program:
                  type: string
                  description: Program to execute
                args:
                  type: array
                  items:
                    type: string
                  description: Program arguments
                cwd:
                  type: string
                  description: Working directory
      responses:
        '200':
          description: Program output
          content:
            application/json:
              schema:
                type: object
                properties:
                  stdout:
                    type: string
                  stderr:
                    type: string
                  exit_code:
                    type: integer
        '400':
          description: Execution error
  /tools:
    get:
      operationId: listTools
      summary: List available tools
      description: Get a list of all available MCP tools
      responses:
        '200':
          description: List of tools
          content:
            application/json:
              schema:
                type: object
                properties:
                  tools:
                    type: array
                    items:
                      type: object
  /health:
    get:
      operationId: healthCheck
      summary: Health check
      description: Check if the server is healthy
      responses:
        '200':
          description: Server is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  server:
                    type: string